third_party/WebKit/LayoutTests/compositing/animation/hidden-composited-expected.txt

Verifies hidden, compositor-animated layers still get composited

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS hiddenLayerIsComposited(documentLayerTree) is true
